labs

+++ 3/src/3.py
@@ -1,7 +1,7 @@
 #3. Арифметика для разных типов
 print('Введите два целых числа ')
-a = int(input("a = "))
-b = int(input("b = "))
+a = float(input("a = "))
+b = float(input("b = "))
 print('a+b= ', a+b)
 print('a-b= ', a-b)
 print('a*b= ', a*b)
